Fix price: change from int (kuruş) to float (TL)
- Pydantic models: price int → float
- Migration: coach_packages.price INT → NUMERIC(10,2)
- Admin panel saves exact TL value, no conversion

+++ app/api/coach/routes.py
@@ -1816,7 +1816,7 @@ class CoachPackageCreate(BaseModel):
     name: str = Field(min_length=1, max_length=120)
     description: Optional[str] = Field(default=None, max_length=2000)
     duration_days: int = Field(gt=0)
-    price: int = Field(ge=0)
+    price: float = Field(ge=0)
     discount_percentage: float = Field(default=0, ge=0, le=100)
     is_active: bool = True
     services: Optional[List[str]] = Field(default=None, description="List of service tags (max 12, each max 40 chars)")
@@ -1832,7 +1832,7 @@ class CoachPackageUpdate(BaseModel):
     name: Optional[str] = Field(default=None, min_length=1, max_length=120)
     description: Optional[str] = Field(default=None, max_length=2000)
     duration_days: Optional[int] = Field(default=None, gt=0)
-    price: Optional[int] = Field(default=None, ge=0)
+    price: Optional[float] = Field(default=None, ge=0)
     discount_percentage: Optional[float] = Field(default=None, ge=0, le=100)
     is_active: Optional[bool] = None
     services: Optional[List[str]] = Field(default=None, description="List of service tags (max 12, each max 40 chars)")
